Impls reverse source code injection
With reverse code injections one can specify the file range of a code
block directly in the source, together with an file wide unique ID
string. Afterwards, this code block can be injected in slides by specifying
the file and the code block ID. Advantage of this technique is that file
ranges can be easier updaten when chaning the code.